refactoring

+++ src/ZeroExpectationRolePlayingSystem.Tests/DiceTestFixture.cs
@@ -16,7 +16,7 @@ namespace ZeroExpectationRolePlayingSystem
         public void Roll(double percentage)
         {
             var dice = new Dice(new Random(0).NextDouble);
-            var values = Enumerable.Range(0, 10000).Select(_ => dice.Roll()).OrderBy(_ => _).ToList();
+            var values = Enumerable.Range(0, 10000).Select(_ => dice.Roll()).OrderBy(_ => (double)_).ToList();
             var samples = (double)values.Count;
             var sampleIndex = (int)((values.Count - 1) * percentage);
             var sample = values[sampleIndex];
